create test for DeleteSizeUseCaseImpl.kt

+++ src/main/kotlin/br/com/hellopizza/api/core/usecase/pizza/impl/DeleteSizeUseCaseImpl.kt
@@ -11,14 +11,16 @@ import java.util.*
 
 @Service
 class DeleteSizeUseCaseImpl(
-        private val sizeGateway: SizeGateway,
-        private val rules: List<DeleteSizeValidationRule>,
-        private val executor: ValidateRuleExecutor,
+    private val sizeGateway: SizeGateway,
+    private val rules: List<DeleteSizeValidationRule>,
+    private val executor: ValidateRuleExecutor,
 ) : DeleteSizeUseCase {
     override suspend fun execute(command: DeleteSizeCommand) {
         val currentState = sizeGateway.findById(command.id)
-                .orElseThrow { SizeNotFoundException(command.id) }
-        executor.validate(rules, Optional.empty(), Optional.of(currentState))
-        sizeGateway.deleteById(command.id)
+            .orElseThrow { SizeNotFoundException(command.id) }
+        val validationResult = executor.validate(rules, Optional.empty(), Optional.of(currentState))
+        if (validationResult.valid) {
+            sizeGateway.deleteById(command.id)
+        }
     }
 }
